fix logo modal routing
